requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asciitree==0.3.3
astroid==3.3.5
attrs==24.2.0
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dill==0.3.9
docker-pycreds==0.4.0
fasteners==0.19
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.67.0
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mccabe==0.7.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numcodecs==0.13.1
numpy==1.24.3
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.2.0
protobuf==5.28.2
psutil==6.1.0
pylint==3.3.1
pyparsing==3.2.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
sentry-sdk==2.17.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.2.2
torch-geometric==2.6.1
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.5
Werkzeug==3.0.4
yarl==1.16.0
zarr==2.18.3
